anno: migrate RelationTriple + DiscontinuousEntity + ScoredRelation confidence to Confidence
RelationTriple.confidence, DiscontinuousEntity.confidence, ScoredRelation.confidence:
f32 -> Confidence. Updated 18 construction sites and 2 read sites across 7 files.

+++ crates/anno-cli/src/cli/commands/extract.rs
@@ -436,7 +436,7 @@ pub fn run(args: ExtractArgs) -> Result<(), CliError> {
 
     // Apply relation confidence threshold (best-effort; non-fatal).
     if args.extract_relations && !relation_triples.is_empty() {
-        relation_triples.retain(|r| r.confidence >= relation_threshold);
+        relation_triples.retain(|r| r.confidence >= relation_threshold as f64);
     }
 
     // Build signals with negation/quantifier annotations, then add to doc.
